chore: merge Web3Names.chain into Did.chain
