+++ packages/did/src/DidLinks/AccountLinks.chain.ts
@@ -32,8 +32,13 @@ import type { PalletDidLookupConnectionRecord } from '@kiltprotocol/augment-api'
 import { ConfigService } from '@kiltprotocol/config'
 
 import { EncodedSignature } from '../Did.utils.js'
-import { Web3Name, web3NameFromChain } from './Web3Names.chain.js'
-import { depositFromChain, toChain, fromChain } from '../Did.chain.js'
+import {
+  depositFromChain,
+  toChain,
+  fromChain,
+  Web3Name,
+  web3NameFromChain,
+} from '../Did.chain.js'
 
 /// A chain-agnostic address, which can be encoded using any network prefix.
 export type SubstrateAddress = KeyringPair['address']
