+++ packages/did/src/DidLinks/index.ts
@@ -4,4 +4,3 @@
  * This source code is licensed under the BSD 4-Clause "Original" license
  * found in the LICENSE file in the root directory of this source tree.
  */
-
