+++ packages/did/src/index.ts
@@ -17,5 +17,4 @@ export * from './DidResolver/index.js'
 export * from './Did.chain.js'
 export * from './Did.utils.js'
 export * from './Did.signature.js'
-export * from './DidLinks/Web3Names.chain.js'
 export * from './DidLinks/AccountLinks.chain.js'
